chatbot/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.3
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.3
googleapis-common-protos==1.63.0
PyPDF2==3.0.1
Werkzeug==3.0.3
zipp==3.18.1
